Change randomized items to use indexed {id, param, mode, player} table
This makes cases where multiple items are in one room much easier to
handle, and enables more items to be multiworld-eligible.

A side effect of this (which needs to be fixed) is that ages minigames
that display the prize in a poof don't have the poof anymore.

I haven't tested every single check to make sure they all work
correctly. I think drops (as in small keys drops and the like) are
broken.